+++ packages/dara-core/js/auth/auth.tsx
@@ -141,7 +141,15 @@ export async function handleAuthErrors(
     if (isAuthenticationError(content?.detail) && !shouldIgnoreError(content?.detail, ignoreErrors)) {
         localStorage.removeItem(getTokenKey());
 
-        const path = toLogin || shouldRedirectToLogin(content.detail) ? '/login' : `/error?code=${res.status}`;
+        // use existing referrer if available in case we were already redirected because of e.g. missing token
+        const queryParams = new URLSearchParams(window.location.search);
+        const referrer =
+            queryParams.get('referrer') ?? encodeURIComponent(window.location.pathname + window.location.search);
+
+        const path =
+            toLogin || shouldRedirectToLogin(content.detail) ?
+                `/login?referrer=${referrer}`
+            :   `/error?code=${res.status}`;
         window.location.href = `${window.dara.base_url}${path}`;
 
         return true;